Improve analytics component
